Change findAccount behavior
Before this commit, the findAccount and the login step used the federation
token as accountId, the federation token was taken from cookies on
X-IO-Federation-Token key.
After this commit, the login and the findAccount use the fiscal_code as
accountId. To do that the findAccount function retrieves the value from
cookies.