FIMS-47 Change client_id to be orgId + serviceId (#42)
* Change how client_id is filled

Before this commit the client_id was generated. After this commit the
client_id is no longer generated, it is filled from organization_id and
service_id in the following format:

<client_id> ::= <organization_id>:<service_id>

* Change the type of clientId on Client structure

Before this commit, the clientId was a string. After this
commit, the clientId became a struct containing organizationId
and serviceId

* Improve test about ClientId